backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf_yasg==1.21.7
django-cors-headers
loadenv
setuptools
pymongo
openai
gunicorn
requests
bs4
nltk
llmlingua
psycopg2-binary